Imported assets and made basic colliders
I imported some assets from the asset store (free) and put them in. I made a tilemap and colliders for a platform and the player. I made layers for background and foreground. I also wrote a movement script but VS code was being weird so it wasn't importing correctly. It might be in this push, it might not. Either way, the movement script is the only thing "not working" here. I reinstalled VS code and will try to fix it later today.